requirements.txt
# Core Flask and API dependencies
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0

# Data processing and ML libraries
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
joblib==1.3.2

# Database and caching
psycopg2-binary==2.9.6  
redis==4.5.4

# Streaming (Kafka)
kafka-python==2.0.2

# Testing
pytest==7.4.3

# Core Python utilities
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3